+++ src/components/Avatar/Avatar.tsx
@@ -2,6 +2,7 @@
  * @author Anna Mikhailova <amikhailova@usetech.com>
  */
 
+import classNames from 'classnames';
 import './Avatar.scss';
 import Empty from '../../assets/static/empty.svg';
 
@@ -10,13 +11,26 @@ export interface IAvatarProps {
     size?: number | string;
     defaultSrc?: string;
     type?: 'circle' | 'square';
+    className?: string;
+    fit?: 'fill' | 'contain' | 'cover' | 'none' | 'scale-down';
 }
 
-export const Avatar = ({ src, size = 38, type = 'square' }: IAvatarProps) => (
+export const Avatar = ({
+    src,
+    size = 38,
+    type = 'square',
+    fit,
+    className,
+}: IAvatarProps) => (
     <img
         width={size}
         height={size}
-        className={`unique-avatar ${type}`}
+        className={classNames(`unique-avatar ${type}`, className)}
+        {...(fit && {
+            style: {
+                objectFit: fit,
+            },
+        })}
         src={src?.length ? src : Empty}
     />
 );